requirements.txt
absl-py==0.14.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
flatbuffers==2.0
gast==0.4.0
google-auth==2.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
idna==3.2
keras-nightly==2.7.0.dev2021092707
Keras-Preprocessing==1.1.2
libclang==11.1.0
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tb-nightly==2.7.0a20210927
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
tf-estimator-nightly==2.7.0.dev2021092408
tf-nightly-gpu==2.8.0.dev20210927
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.1
wrapt==1.12.1
